refactor: MenuOptionDto, MenuSelectionDto 클래스의 패키지 이동
- store 모듈의 테스트 디렉토리 -> common모듈의 store.business.internal.dto
